src/python/server/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.0
Werkzeug==2.2.2
zipp==3.11.0
tflite-runtime==2.11.0
pillow==9.3.0
